test: move BunTranspiledModule debugger probe to its own file
inspect.test.ts has pre-existing localhost-based websocket cases that are
environment-sensitive (and the file is already [ASAN] [TIMEOUT] in
test/expectations.txt); running it end to end fails on those unrelated cases.
The new cases live in debugger-buntranspiledmodule.test.ts so the regression
guard runs clean on its own.
